+++ jina/checker.py
@@ -61,7 +61,7 @@ class NetworkChecker:
             total_success = 0
             for j in range(args.retries):
                 with TimeContext(f'ping {ctrl_addr} at {j} round', default_logger) as tc:
-                    r = send_ctrl_message(ctrl_addr, 'STATUS', timeout=args.timeout_ctrl)
+                    r = send_ctrl_message(ctrl_addr, 'STATUS', timeout=args.timeout)
                     if not r:
                         default_logger.warning('not responding, retry (%d/%d) in 1s' % (j + 1, args.retries))
                     else:
